instead of using a complex decorator, make getting the environment before calling items more simpler.
we now have runner (DockLightServiceRunner). This runner will inject the needed operations call from the IDockerclient to the service. This way, the service can focus on what we want from te service (like getting container data and such)

the runner will get the api location from the created environment and will use that inffomraiton to constrcut the IDockerClient,. Afterwards we check in our naive cache if we have that service and return it.

The cache is key value with the key being the service interface and the value being a factory, that get's constructed when needed. Should theoreticall only be called once per scope.

it atleast removes the complex decorator call