Revamp how we get docker data + fixes (#14)
* add mui lab

* update frontend for some changes to also see our data getting fetched (health statys null tho )

* move and add extra models

* remove not needed controller

* add new queries and handelrs for containers.

* add extra json options for serializing

* instead of using a complex decorator, make getting the environment before calling items more simpler.

we now have runner (DockLightServiceRunner). This runner will inject the needed operations call from the IDockerclient to the service. This way, the service can focus on what we want from te service (like getting container data and such)

the runner will get the api location from the created environment and will use that inffomraiton to constrcut the IDockerClient,. Afterwards we check in our naive cache if we have that service and return it.

The cache is key value with the key being the service interface and the value being a factory, that get's constructed when needed. Should theoreticall only be called once per scope.

it atleast removes the complex decorator call

* fix red squisqly thingies

* remove forgotten usings

* custom fonts, naming

* naming for fixing filename

* fixed filename

* add extra components for table row (container detail) and a Date time ago widget to show the date as "about xx hours ago"

* update for naming NordTheme

* support for fonts

* fixes, show some details

* make sure enums are being sent as strings and (lowercase) and not as numbers.

* Add helper for getting services

* rename to factory, as it;s a factory.

* also rename to factory and add a memory cache. several other fixes

* fixes, use helper methods

* add empty services and interfaces for images and system

* Add Memorycache for every service

* Use inheritance for helpermethod instead of extension methods

* small fixes

* fixes after code review