refactor startup features, make them more independabl.
The previous attempt assummed (thnx AI) that we wanted to confiugre services and application in one app. But that is not the case.

It's now more explicit which startup features we have, making it more easier to add a new startup feature, if ever needed.

most of the time a new component will be added.