Refactor CreateDockLightEnvironmentHandler
Refactored `CreateDockLightEnvironmentHandler` to use `IDockLightEnvironmentService` instead of `IDatabaseWriteRepository`. Updated the `Handle` method to delegate environment creation to the service. Changed the class visibility to `internal sealed`.

Added `ISharedDockLightEnvironmentService` to the service collection in `DockLightEnvironmentServiceComponent`. Updated the interface to inherit from `IViixsBaseService` and added a method for creating environments.

Adjusted namespaces in relevant files to reflect the new dependencies. Added a `TODO` comment for WebSocket feedback implementation.